RMA working 1 Formatted all sources
	Have the particle_rma_comm and worker_rma working
	Used clang-format -style=LLVM on src src_test and include

+++ src/yaml_loader.cpp
@@ -19,7 +19,7 @@ YamlLoader::YamlLoader(std::string filepath) {
       s = s.substr(0, s.length() - 1);
     }
     int n = s.find(":");
-    if (n == std::string::npos) {
+    if (n == (int)std::string::npos) {
       fprintf(stderr, "Yaml File %s was not correctly formatted.\n",
               filepath.c_str());
       exit(1);
